% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angular_loglik.R
\name{angular_loglik}
\alias{angular_loglik}
\title{Run adaptive MH algorithm for angular parameters}
\usage{
angular_loglik(angles, dim, posterior_params, gauge_type)
}
\description{
Run adaptive MH algorithm for angular parameters
}
