% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angular_mcmc.R
\name{angular_mcmc}
\alias{angular_mcmc}
\title{Run adaptive MH algorithm for angular parameters}
\usage{
angular_mcmc(
  angles,
  dim,
  starting_theta,
  gauge_type,
  prop_Sigma_ = NULL,
  n_updates = 10000,
  update_freq = 100,
  n_burnin = 0,
  n_thin = 1,
  adapt_cov = FALSE,
  r_opt = 0.234,
  c0 = 10,
  c1 = 0.8,
  K = 10
)
}
\description{
Run adaptive MH algorithm for angular parameters
}
