% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaugeDependence-package.R
\name{asym_log_gauge}
\alias{asym_log_gauge}
\title{Asymmetric logistic gauge function}
\arguments{
\item{w1, w2}{Numeric vectors of angular coordinates (with \code{w1 + w2 = 1}).}

\item{dep}{Numeric dependence parameter in (0, 1].}
}
\value{
Numeric vector of gauge function values.
}
\description{
Asymmetric logistic gauge function
}
