% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaugeDependence-package.R
\name{dirichlet_gauge}
\alias{dirichlet_gauge}
\title{Dirichlet gauge function}
\arguments{
\item{w1, w2}{Numeric vectors of angular coordinates (with \code{w1 + w2 = 1}).}

\item{dep}{Numeric vector of two positive dependence parameters,
\code{c(theta1, theta2)}.}
}
\value{
Numeric vector of gauge function values.
}
\description{
Dirichlet gauge function
}
