\name{gaugeDep-package}
\alias{gaugeDep-package}
\alias{gaugeDep}
\docType{package}
\title{\packageTitle{gaugeDep}}
\description{\packageDescription{gaugeDep}}
\section{Package Content}{\packageIndices{gaugeDep}}
\author{\packageAuthor{gaugeDep}}
\section{Maintainer}{\packageMaintainer{gaugeDep}}
\keyword{package}
