% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaugeDependence-package.R
\docType{package}
\name{gaugeDependence-package}
\alias{gaugeDependence}
\alias{gaugeDependence-package}
\title{gaugeDependence: Bayesian MCMC for Bivariate Extremes via Gauge Functions}
\description{
Adaptive Metropolis-Hastings MCMC samplers for fitting bivariate extreme
value models using gauge function dependence structures. The joint
distribution of a bivariate extreme is decomposed into radial (R) and
angular (W) components: radii above a threshold are modelled as
Gamma(2, g(W)), where g is a gauge function; angles are modelled via a
star-shaped angular density. Six parametric gauge families are implemented,
with truncated and censored likelihood options for the radial component.
}
\section{Main functions}{

\describe{
  \item{\code{\link{radial_mcmc}}}{Fit the radial model via adaptive MH MCMC.}
  \item{\code{\link{angular_mcmc}}}{Fit the star-shaped angular model via adaptive MH MCMC.}
  \item{\code{\link{radial_loglik}}}{Posterior pointwise log-likelihood for the radial model.}
  \item{\code{\link{angular_loglik}}}{Posterior pointwise log-likelihood for the angular model.}
  \item{\code{\link{reject_sampling}}}{Sample angles from the star-shaped density.}
}
}

\section{Gauge functions}{

The six gauge function families are also exported and can be evaluated
directly: \code{\link{gauss_gauge}}, \code{\link{logistic_gauge}},
\code{\link{inv_log_gauge}}, \code{\link{asym_log_gauge}},
\code{\link{dirichlet_gauge}}, and \code{\link{rectangular_gauge}}.
}

\author{
\strong{Maintainer}: Elizabeth Lawler \email{eslawler@gmail.com} (\href{https://orcid.org/0009-0004-4491-4733}{ORCID})

}
\keyword{internal}
