% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaugeDependence-package.R
\name{inv_log_gauge}
\alias{inv_log_gauge}
\title{Inverted logistic gauge function}
\arguments{
\item{w1, w2}{Numeric vectors of angular coordinates (with \code{w1 + w2 = 1}).}

\item{dep}{Numeric dependence parameter in (0, 1].}
}
\value{
Numeric vector of gauge function values.
}
\description{
Inverted logistic gauge function
}
