% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radial_angular_mcmc.R
\name{radial_angular_mcmc}
\alias{radial_angular_mcmc}
\title{Run adaptive MH algorithm for radial-angular parameters}
\usage{
radial_angular_mcmc(
  radii,
  angles,
  threshold,
  sum_term,
  sqrt_term,
  dim,
  starting_theta_rad,
  starting_theta_ang,
  rad_gauge_type,
  ang_gauge_type,
  prop_Sigma_radii_ = NULL,
  prop_Sigma_angles_ = NULL,
  n_updates = 10000,
  update_freq = 100,
  n_burnin = 0,
  n_thin = 1,
  adapt_cov = FALSE,
  r_opt = 0.234,
  c0 = 10,
  c1 = 0.8,
  K = 10
)
}
\description{
Run adaptive MH algorithm for radial-angular parameters
}
