% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radial_loglik.R, R/radial_mcmc.R
\name{radial_mcmc}
\alias{radial_mcmc}
\title{Run adaptive MH algorithm for radial parameters}
\usage{
radial_mcmc(
  radii,
  threshold,
  angles,
  starting_theta,
  likelihood_type,
  gauge_type,
  prop_Sigma_ = NULL,
  n_updates = 10000,
  update_freq = 100,
  n_burnin = 0,
  n_thin = 1,
  adapt_cov = FALSE,
  r_opt = 0.234,
  c0 = 10,
  c1 = 0.8,
  K = 10
)

radial_mcmc(
  radii,
  threshold,
  angles,
  starting_theta,
  likelihood_type,
  gauge_type,
  prop_Sigma_ = NULL,
  n_updates = 10000,
  update_freq = 100,
  n_burnin = 0,
  n_thin = 1,
  adapt_cov = FALSE,
  r_opt = 0.234,
  c0 = 10,
  c1 = 0.8,
  K = 10
)
}
\description{
Run adaptive MH algorithm for radial parameters

Run adaptive MH algorithm for radial parameters
}
