\name{gaugeDependence-package}
\alias{gaugeDependence-package}
\alias{gaugeDependence}
\docType{package}
\title{\packageTitle{gaugeDependence}}
\description{\packageDescription{gaugeDependence}}
\section{Package Content}{\packageIndices{gaugeDependence}}
\author{\packageAuthor{gaugeDependence}}
\section{Maintainer}{\packageMaintainer{gaugeDependence}}
\keyword{package}
